Moving gdpr to `talk-plugin-profile-data`
